Working concept of a WebDAV server
The server runs a a Windows service. Upon install, it automatically tries to get or create a valid server Certificate from the store. It then binds the certificate to the end point where the HttpListener will be listening on.

The WebDAV server provides only basic authentication. All permissions on the file system are acquired from the user that owns the service process.